backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ── Config ────────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0

# ── HTTP clients ─────────────────────────────────────────────────────────────
httpx==0.27.0
aiohttp==3.9.5

# ── Data sources ─────────────────────────────────────────────────────────────
amadeus==9.0.0
duffel-api==0.6.0

# ── AI ────────────────────────────────────────────────────────────────────────
anthropic==0.28.0

# ── Alerts ───────────────────────────────────────────────────────────────────
twilio==9.1.0

# ── Utilities ────────────────────────────────────────────────────────────────
redis[asyncio]==5.0.4
python-dotenv==1.0.1
structlog==24.1.0
numpy==1.26.4
orjson==3.10.3
